requirements.txt
# ── Core orchestration ────────────────────────────────────
langgraph>=0.2
langchain-core>=0.3
langchain-openai>=0.2
langchain-anthropic>=0.3
langchain-google-genai>=2.0
langchain-ollama>=0.2
langchain-deepseek>=0.1       # DeepSeek
# xAI (Grok) uses the OpenAI-compatible client via langchain-openai

# ── Crypto exchange data ──────────────────────────────────
ccxt[pro]>=4.3
pybit>=5.14.0                   # Bybit SDK

# ── Data processing ───────────────────────────────────────
pandas>=2.2
numpy>=2.0
polars>=0.20
pandas-ta>=0.3.14b

# ── Orderflow & Volume Profile ────────────────────────────
# Built in-house: data/orderflow.py + data/volume_profile.py
# (orderflow PyPI package is incompatible with Python 3.12)

# ── HTTP clients ──────────────────────────────────────────
requests>=2.31
aiohttp>=3.9
httpx>=0.27

# ── Database ──────────────────────────────────────────────
psycopg2-binary>=2.9
asyncpg>=0.29
redis>=5.0
chromadb>=0.5
sqlalchemy>=2.0
alembic>=1.13

# ── API server ────────────────────────────────────────────
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
pydantic>=2.7
pydantic-settings>=2.3

# ── Backtesting ───────────────────────────────────────────
vectorbt>=0.26

# ── Utilities ─────────────────────────────────────────────
loguru>=0.7
schedule>=1.2
pytz>=2024.1
